+++ src/i2c_driver.h
@@ -87,7 +87,7 @@ public:
     // Set 'send_stop' to true if this is the last transfer in the transaction.
     // Set 'send_stop' to false if are going to make another transfer.
     // Call finished() to see if the call has finished.
-    virtual void write_async(uint16_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) = 0;
+    virtual void write_async(uint8_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) = 0;
 
     // Reads the specified number of bytes and copies them into the supplied buffer.
     // The caller must not modify the buffer until the read is complete.
@@ -95,13 +95,13 @@ public:
     // Set 'send_stop' to true if this is the last transfer in the transaction.
     // Set 'send_stop' to false if are going to make another transfer.
     // Call finished() to see if the call has finished.
-    virtual void read_async(uint16_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) = 0;
+    virtual void read_async(uint8_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) = 0;
 };
 
 class I2CSlave : public I2CDriver {
 public:
     // Start listening to the master on the given address. Makes the slave visible on the bus.
-    virtual void listen(uint16_t address) = 0;
+    virtual void listen(uint8_t address) = 0;
 
     // Sets a callback to be called by the ISR each time
     // the slave receives a block of data from the master.
