+++ src/imx_rt1060/imx_rt1060_i2c_driver.cpp
@@ -146,7 +146,7 @@ size_t IMX_RT1060_I2CMaster::get_bytes_transferred() {
     return buff.get_bytes_transferred();
 }
 
-void IMX_RT1060_I2CMaster::write_async(uint16_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) {
+void IMX_RT1060_I2CMaster::write_async(uint8_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) {
     if (!start(address, MASTER_WRITE)) {
         return;
     }
@@ -163,7 +163,7 @@ void IMX_RT1060_I2CMaster::write_async(uint16_t address, uint8_t* buffer, size_t
     port->MIER |= LPI2C_MIER_TDIE;
 }
 
-void IMX_RT1060_I2CMaster::read_async(uint16_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) {
+void IMX_RT1060_I2CMaster::read_async(uint8_t address, uint8_t* buffer, size_t num_bytes, bool send_stop) {
     if (num_bytes > MAX_MASTER_READ_LENGTH) {
         _error = I2CError::invalid_request;
         return;
@@ -305,7 +305,7 @@ inline void IMX_RT1060_I2CMaster::clear_all_msr_flags() {
                   LPI2C_MSR_EPF | LPI2C_MSR_RDF | LPI2C_MSR_TDF);
 }
 
-bool IMX_RT1060_I2CMaster::start(uint16_t address, uint32_t direction) {
+bool IMX_RT1060_I2CMaster::start(uint8_t address, uint32_t direction) {
     if (!finished()) {
         // We haven't completed the previous transaction yet
         #ifdef DEBUG_I2C
@@ -408,7 +408,7 @@ void IMX_RT1060_I2CMaster::set_clock(uint32_t frequency) {
     port->MCCR1 = port->MCCR0;
 }
 
-void IMX_RT1060_I2CSlave::listen(uint16_t address) {
+void IMX_RT1060_I2CSlave::listen(uint8_t address) {
     // Make sure slave mode is disabled before configuring it.
     stop_listening();
